Keep dataset window on-screen when it opens
I considered centering it in the body, like the bookmarks window, but
without an overlay behind it it did not look great. consider this a
first step; I'm also working on altering the layout of the drop so that
it is easier to work with and easier to support new data load methods.